[JAX] Renamed config files and paths to match naming conventions.
