+++ requirements.txt
@@ -22,5 +22,5 @@ openpyxl
 netmiko
 jinja2
 
-# Headless plotting for training loss-curve PNGs (Agg backend; see cvs/lib/training/jax/utils/loss_curve.py)
+# Headless plotting for training loss-curve PNGs (Agg backend; see cvs/lib/training/jaxmaxtext/utils/loss_curve.py)
 matplotlib
